Breakout topology package options for readability
